Refactor code to apply user feedback form on both mainboard and landing page
