+++ src/components/MainBoard.js
@@ -5,8 +5,7 @@ import ChatBox from './ChatBox';
 import PlotBox from './PlotBox';
 import { triggersPlotUpdate } from '../utils/chatSideEffects';
 import { updatePlotState } from '../utils/updatePlotState';
-import { FloatButton } from 'antd';
-import { QuestionCircleOutlined } from '@ant-design/icons';
+import FeedbackForm from './FeedbackForm';
 
 const MainBoard = () => {
   const location = useLocation();
@@ -47,14 +46,9 @@ const MainBoard = () => {
           <div style={{ flex: 1, overflow: 'auto'}}>
             {plotState && <PlotBox state={plotState} />}
           </div>
-          <FloatButton
-            icon={<QuestionCircleOutlined />}
-            type="default"
-            style={{
-              right: 30,
-            }}
-          />
+         
       </div>
+      <FeedbackForm/>
     </ChatProvider>
   );
 };
